+++ java/src/de/funroll_loops/oscar/nc/StringTable.java
@@ -1,9 +1,9 @@
 package de.funroll_loops.oscar.nc;
 
 public class StringTable extends Base implements BaseInterface {
-	public native int create();
-	public native void destroy(int ptr);
+	public native long create();
+	public native void destroy(long ptr);
 	
-	public native int size(int ptr);
-	public native String at(int ptr, int pos);
+	public native int size(long ptr);
+	public native String at(long ptr, int pos);
 }
